Move parallelism to WorkerNode.Property, update related structs and references

+++ src/meta/src/stream/stream_manager.rs
@@ -627,10 +627,7 @@ impl GlobalStreamManager {
             .collect::<BTreeSet<_>>();
 
         // Check if the provided parallelism is valid.
-        let available_parallelism = worker_nodes
-            .iter()
-            .map(|w| w.parallelism as usize)
-            .sum::<usize>();
+        let available_parallelism = worker_nodes.iter().map(|w| w.parallelism()).sum::<usize>();
         let max_parallelism = self
             .metadata_manager
             .get_job_max_parallelism(table_id)
